Generalise regex url matcher
